# code/go/0chain.net/blobbercore/openapi/blobber.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "blobber.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v2/allocation": {
      "get": {
        "operationId": "Blobber_GetAllocation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAllocationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/connection/commit/{allocation}": {
      "post": {
        "operationId": "Blobber_Commit",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CommitResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CommitRequest"
            }
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/attributes/{allocation}": {
      "post": {
        "operationId": "Blobber_UpdateObjectAttributes",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateObjectAttributesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateObjectAttributesRequest"
            }
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/calculatehash/{allocation}": {
      "post": {
        "operationId": "Blobber_CalculateHash",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CalculateHashResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CalculateHashRequest"
            }
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/collaborator/{allocation}": {
      "post": {
        "operationId": "Blobber_Collaborator",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CollaboratorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CollaboratorRequest"
            }
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/commitmetatxn/{allocation}": {
      "post": {
        "operationId": "Blobber_CommitMetaTxn",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CommitMetaTxnResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CommitMetaTxnRequest"
            }
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/copy/{allocation}": {
      "post": {
        "operationId": "Blobber_CopyObject",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CopyObjectResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CopyObjectRequest"
            }
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/download/{allocation}": {
      "post": {
        "operationId": "Blobber_DownloadFile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DownloadFileResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1DownloadFileRequest"
            }
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/list/{allocation}": {
      "get": {
        "operationId": "Blobber_ListEntities",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListEntitiesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pathHash",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "authToken",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/meta/{allocation}": {
      "get": {
        "operationId": "Blobber_GetFileMetaData",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetFileMetaDataResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pathHash",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "authToken",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/objectpath/{allocation}": {
      "get": {
        "operationId": "Blobber_GetObjectPath",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetObjectPathResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Path",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "BlockNum",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/objecttree/{allocation}": {
      "get": {
        "operationId": "Blobber_GetObjectTree",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetObjectTreeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/referencepath/{allocation}": {
      "get": {
        "operationId": "Blobber_GetReferencePath",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetReferencePathResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Paths",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "Path",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/rename/{allocation}": {
      "post": {
        "operationId": "Blobber_RenameObject",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RenameObjectResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RenameObjectRequest"
            }
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    },
    "/v2/file/stats/{allocation}": {
      "get": {
        "operationId": "Blobber_GetFileStats",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetFileStatsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "allocation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pathHash",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Blobber"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1CalculateHashRequest": {
      "type": "object",
      "properties": {
        "allocation": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "paths": {
          "type": "string"
        }
      }
    },
    "v1CalculateHashResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "v1Collaborator": {
      "type": "object",
      "properties": {
        "RefId": {
          "type": "string",
          "format": "int64"
        },
        "ClientId": {
          "type": "string"
        },
        "CreatedAt": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1CollaboratorRequest": {
      "type": "object",
      "properties": {
        "allocation": {
          "type": "string"
        },
        "collabId": {
          "type": "string"
        },
        "method": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "pathHash": {
          "type": "string"
        }
      }
    },
    "v1CollaboratorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "Collaborators": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Collaborator"
          }
        }
      }
    },
...
    "v1CommitMetaTxnRequest": {
      "type": "object",
      "properties": {
        "path": {
          "type": "string"
        },
        "pathHash": {
          "type": "string"
        },
        "authToken": {
          "type": "string"
        },
        "allocation": {
          "type": "string"
        },
        "txnId": {
          "type": "string"
        }
      }
    },
    "v1CommitMetaTxnResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "v1CommitRequest": {
      "type": "object",
      "properties": {
        "allocation": {
          "type": "string"
        },
        "connectionId": {
          "type": "string"
        },
        "writeMarker": {
          "type": "string"
        }
      }
    },
    "v1CommitResponse": {
      "type": "object",
      "properties": {
        "allocationRoot": {
          "type": "string"
        },
        "writeMarker": {
          "$ref": "#/definitions/v1WriteMarker"
        },
        "errorMessage": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      }
    },
    "v1CopyObjectRequest": {
      "type": "object",
      "properties": {
        "allocation": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "pathHash": {
          "type": "string"
        },
        "connectionId": {
          "type": "string"
        },
        "dest": {
          "type": "string"
        }
      }
    },
    "v1CopyObjectResponse": {
      "type": "object",
      "properties": {
        "filename": {
          "type": "string"
        },
        "size": {
          "type": "string",
          "format": "int64"
        },
        "contentHash": {
          "type": "string"
        },
        "merkleRoot": {
          "type": "string"
        },
        "uploadLength": {
          "type": "string",
          "format": "int64",
          "description": "UploadLength indicates the size of the entire upload in bytes. The value MUST be a non-negative integer."
        },
        "uploadOffset": {
          "type": "string",
          "format": "int64",
          "description": "Upload-Offset indicates a byte offset within a resource. The value MUST be a non-negative integer."
        }
      }
    },
...
    "v1DownloadFileRequest": {
      "type": "object",
      "properties": {
        "allocation": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "pathHash": {
          "type": "string"
        },
        "rxPay": {
          "type": "string"
        },
        "blockNum": {
          "type": "string"
        },
        "numBlocks": {
          "type": "string"
        },
        "readMarker": {
          "type": "string"
        },
        "authToken": {
          "type": "string"
        },
        "content": {
          "type": "string"
        }
      }
    },
    "v1DownloadFileResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "data": {
          "type": "string",
          "format": "byte"
        },
        "allocationId": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "latestRm": {
          "$ref": "#/definitions/v1ReadMaker"
        }
      }
    },
...
    "v1GetAllocationResponse": {
      "type": "object",
      "properties": {
        "allocation": {
          "$ref": "#/definitions/v1Allocation"
        }
      }
    },
    "v1GetFileMetaDataResponse": {
      "type": "object",
      "properties": {
        "MetaData": {
          "$ref": "#/definitions/v1FileRef"
        },
        "Collaborators": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Collaborator"
          }
        }
      }
    },
    "v1GetFileStatsResponse": {
      "type": "object",
      "properties": {
        "MetaData": {
          "$ref": "#/definitions/v1FileRef"
        },
        "Stats": {
          "$ref": "#/definitions/v1FileStats"
        }
      }
    },
    "v1GetObjectPathResponse": {
      "type": "object",
      "properties": {
        "ObjectPath": {
          "$ref": "#/definitions/v1ObjectPath"
        },
        "LatestWriteMarker": {
          "$ref": "#/definitions/v1WriteMarker"
        }
      }
    },
    "v1GetObjectTreeResponse": {
      "type": "object",
      "properties": {
        "ReferencePath": {
          "$ref": "#/definitions/v1ReferencePath"
        },
        "LatestWM": {
          "$ref": "#/definitions/v1WriteMarker"
        }
      }
    },
    "v1GetReferencePathResponse": {
      "type": "object",
      "properties": {
        "ReferencePath": {
          "$ref": "#/definitions/v1ReferencePath"
        },
        "LatestWM": {
          "$ref": "#/definitions/v1WriteMarker"
        }
      }
    },
    "v1ListEntitiesResponse": {
      "type": "object",
      "properties": {
        "AllocationRoot": {
          "type": "string"
        },
        "MetaData": {
          "$ref": "#/definitions/v1FileRef"
        },
        "Entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1FileRef"
          }
        }
      }
    },
...
    "v1RenameObjectRequest": {
      "type": "object",
      "properties": {
        "allocation": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "pathHash": {
          "type": "string"
        },
        "connectionId": {
          "type": "string"
        },
        "newName": {
          "type": "string"
        }
      }
    },
    "v1RenameObjectResponse": {
      "type": "object",
      "properties": {
        "filename": {
          "type": "string"
        },
        "size": {
          "type": "string",
          "format": "int64"
        },
        "contentHash": {
          "type": "string"
        },
        "merkleRoot": {
          "type": "string"
        },
        "uploadLength": {
          "type": "string",
          "format": "int64",
          "description": "UploadLength indicates the size of the entire upload in bytes. The value MUST be a non-negative integer."
        },
        "uploadOffset": {
          "type": "string",
          "format": "int64",
          "description": "Upload-Offset indicates a byte offset within a resource. The value MUST be a non-negative integer."
        }
      }
    },
...
    "v1UpdateObjectAttributesRequest": {
      "type": "object",
      "properties": {
        "allocation": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "pathHash": {
          "type": "string"
        },
        "connectionId": {
          "type": "string"
        },
        "attributes": {
          "type": "string"
        }
      }
    },
    "v1UpdateObjectAttributesResponse": {
      "type": "object",
      "properties": {
        "whoPaysForReads": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "v1WriteMarker": {
      "type": "object",
      "properties": {
        "AllocationRoot": {
          "type": "string"
        },
        "PreviousAllocationRoot": {
          "type": "string"
        },
        "AllocationID": {
          "type": "string"
        },
        "Size": {
          "type": "string",
          "format": "int64"
        },
        "BlobberID": {
          "type": "string"
        },
        "Timestamp": {
          "type": "string",
          "format": "int64"
        },
        "ClientID": {
          "type": "string"
        },
        "Signature": {
          "type": "string"
        }
      }
    }
  }
}